addinguser: don't say "log in as lfs user"
"log in" is a specifically defined action on UNIces.  When you log in
as a user, an entry will be added into /run/utmp and /var/log/wtmp.
But when you switch to lfs with "su", this won't happen.

"man su" also describes "--login" as:

    Provide an environment similar to what the user would expect had
    the user logged in directly.

It's "similar to" logged in, not "really" logged in.

So, don't say "log in" or the users may assume they must really log in
as lfs in a VT or DM.