Emergency kernel update
Update to Linux-5.19.13

If you are using a laptop with an Intel GPU, it's imperative that you
update immediately if you are running Linux-5.19.12. Failure to upgrade
may result in permanent damage to the LCD display on your laptop.

The root cause of this is improper backporting of bugfixes for the i915
DRM driver in the kernel.